+++ HISTORY.rst
@@ -14,8 +14,8 @@ FUTURE (UN-RELEASED)
 
 - Moving upload of wheel to use twine instead of setup.py. See the following link for why: https://packaging.python.org/distributing/#wheels
 
-2.11.5 (UN-RELEASED)
---------------------
+2.11.5 (2017-05-01)
+-------------------
 
 **New Features:**